Split IMusicLibrary into an array of Independant Services/Repos
