+++ Dukebox.Library/Interfaces/IMusicLibraryCacheService.cs
@@ -1,0 +1,14 @@
+﻿using System.Collections.Generic;
+using Dukebox.Library.Model;
+
+namespace Dukebox.Library.Interfaces
+{
+    public interface IMusicLibraryCacheService
+    {
+        List<Album> OrderedAlbums { get; }
+        List<Artist> OrderedArtists { get; }
+        List<Playlist> OrderedPlaylists { get; }
+        List<string> FilesCache { get; }
+        void RefreshCaches();
+    }
+}
