+++ Dukebox.Library/Interfaces/IMusicLibraryUpdateService.cs
@@ -1,0 +1,11 @@
+﻿using System.Threading.Tasks;
+using Dukebox.Library.Model;
+
+namespace Dukebox.Library.Interfaces
+{
+    public interface IMusicLibraryUpdateService
+    {
+        Task RemoveTrack(ITrack track);
+        Task SaveSongChanges(Song song);
+    }
+}
