+++ Dukebox.Library/Interfaces/IPlaylistGeneratorService.cs
@@ -1,0 +1,9 @@
+﻿using Dukebox.Library.Model;
+
+namespace Dukebox.Library.Interfaces
+{
+    public interface IPlaylistGeneratorService
+    {
+        Playlist GetPlaylistFromFile(string playlistFile);
+    }
+}
